Total:151(unchanged)
I resolved this this problem with recursive method, and time complexity is O(n), which n is the node number of array.
Algorithm is simple, just calculate the middle of array, since that is sorted one. And recursive calling the function to handle the left and right subtrees.
Got Accepted solution.

Change Contents:
    None

Signed-off-by: robearro@gmail.com